backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-jalali-date==1.1.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
djangorestframework==3.15.0
jdatetime==4.1.1
pillow==10.2.0
setuptools==69.2.0
sqlparse==0.4.4
